A better solution for language setting
Thanks to yiidoc/yii2-redactor, made a better solution for language setting through AssetBundle properties manipulation. Registering an asset bundle makes a View file know what scripts and styles to register and it happens in the end of the Application workflow, during page rendering process, so it is possible to add scripts and styles _after_ registering an asset bundle.

- ImperaviRedactorAsset
  - [REMOVED] Unnecessary init() method
- Widget
  - [REMOVED] Unnecessary setLanguageOption()
  - [ADDED] registerRedactorAsset() for registering the asset bundle to a protected variable
  - [CHANGE] Asset Bundle registration is moved from run() to the method above.
  - [ADDED] getAssetBundle() for asset bundle properties manipulation
  - [ADDED] Language javascript file addition to js[] array of the registered asset bundle in registerClienScript().